consolidate tests and utils
